Separated graph class to its own file. Renamed to StatsGraph. Reused DoubleFormatter to remove trailing decimal zeroes
